+++ packages/angular_devkit/build_angular/builders.json
@@ -1,6 +1,11 @@
 {
   "$schema": "../architect/src/builders-schema.json",
   "builders": {
+    "application": {
+      "implementation": "./src/builders/application",
+      "schema": "./src/builders/application/schema.json",
+      "description": "Build an application."
+    },
     "app-shell": {
       "implementation": "./src/builders/app-shell",
       "schema": "./src/builders/app-shell/schema.json",
