+++ packages/angular_devkit/build_angular/src/builders/application/tests/options/subresource-integrity_spec.ts
@@ -7,10 +7,10 @@
  */
 
 import { logging } from '@angular-devkit/core';
-import { buildEsbuildBrowser } from '../../index';
-import { BASE_OPTIONS, BROWSER_BUILDER_INFO, describeBuilder } from '../setup';
+import { buildApplication } from '../../index';
+import { APPLICATION_BUILDER_INFO, BASE_OPTIONS, describeBuilder } from '../setup';
 
-describeBuilder(buildEsbuildBrowser, BROWSER_BUILDER_INFO, (harness) => {
+describeBuilder(buildApplication, APPLICATION_BUILDER_INFO, (harness) => {
   describe('Option: "subresourceIntegrity"', () => {
     it(`does not add integrity attribute when not present`, async () => {
       harness.useTarget('build', {
